+++ extensions/broker-server/src/main/java/de/sovity/edc/ext/brokerserver/dao/pages/catalog/models/CatalogQueryFilter.java
@@ -14,11 +14,11 @@
 
 package de.sovity.edc.ext.brokerserver.dao.pages.catalog.models;
 
-import java.util.List;
+import lombok.NonNull;
 
 public record CatalogQueryFilter(
-        String searchQuery,
-        List<CatalogQuerySelectedFilterQuery> selectedFilters
+        @NonNull String name,
+        @NonNull AvailableFilterValuesQuery valueQuery,
+        CatalogQuerySelectedFilterQuery queryFilterClauseOrNull
 ) {
-
 }